refactor(agw): separate dhcp helper cli (#14635)
* refactor(agw): separate dhcp helper cli

This PR separates the DHCP helper CLI into its own folder and creates a separate release for it. The setup is to build a deb artifact with bazel, which is installed alongside Magma. The CLI tool can then be used by mobilityd in the dev VM, deb VM, and the containerized AGW.


* fix(agw): Fix docker build with new script


* chore(agw): add license file for dhcp_helper_cli


* chore(agw): add readme for dhcp_helper_cli


* refactor(agw): Removed an f from a string


* refactor(agw): Only look at last line from the helper script


* fix(agw): install dhcp_helper_cli on dev vm during provisioning


* fix(agw): enable rerun of sudo tests with bazel


* chore(agw): adapt bazel built dhcp cli deb metadata


* fix(agw): fix dhcp cli response parsing


* refactor(agw): Fix some linting issues


* refactor(agw): Fixed json decoding issue


* refactor(agw): Changed filename


* refactor(agw): Changed some package metadata


* chore(agw): Extract new DHCP tool into a new package

WIP


* refactor(agw): Refactored build-magma.sh


* refactor(agw): Fixed a path


* fix(agw): make linter happy


* chore(agw): change comment for dhcp cli usage


* chore(agw): remove dangling comment


* chore(agw): add min version for dhcp cli in make build


* feat(agw): add license field to bazel dhcp cli deb


* fix(agw): remove dhcp helper cli from make build deb


* chore(agw): add license file to dhcp cli deb package


* fix(agw): format changes


* chore(agw): Use consistent name for DHCP tool license


* chore(agw): Formatting of BUILD.bazel
