+++ lte/gateway/docker/services/python/Dockerfile
@@ -76,6 +76,7 @@ COPY \
     $MAGMA_ROOT/lte/gateway/python/
 COPY ./lte/gateway/python/load_tests $MAGMA_ROOT/lte/gateway/python/load_tests
 COPY ./lte/gateway/python/scripts $MAGMA_ROOT/lte/gateway/python/scripts
+COPY ./lte/gateway/python/dhcp_helper_cli $MAGMA_ROOT/lte/gateway/python/dhcp_helper_cli
 COPY ./lte/gateway/python/magma $MAGMA_ROOT/lte/gateway/python/magma
 COPY ./orc8r/gateway/python $MAGMA_ROOT/orc8r/gateway/python
 
@@ -126,6 +127,7 @@ RUN apt-get update && apt-get install -y \
   python3-eventlet \
   python3-pystemd \
   python3-jinja2 \
+  python3-scapy \
   redis-server \
   sudo \
   virtualenv \
@@ -156,6 +158,7 @@ RUN apt-get update && apt-get install -y \
 COPY \
     ./orc8r/gateway/python/scripts/ \
     ./lte/gateway/python/scripts/ \
+    ./lte/gateway/python/dhcp_helper_cli/dhcp_helper_cli.py \
     /usr/local/bin/
 COPY ./lte/gateway/configs/templates /etc/magma/templates/
 COPY ./lte/gateway/deploy/roles/magma/files/set_irq_affinity /usr/local/bin/set_irq_affinity
@@ -165,6 +168,9 @@ COPY ./orc8r/gateway/python/magma /magma/orc8r/gateway/python/magma
 
 COPY --from=builder /build /build
 COPY --from=builder /var/tmp/python3-aioeventlet*.deb /var/tmp/
+
+# The Python scripts in /build/bin are broken inside the container
+RUN rm -f /build/bin/dhcp_helper_cli.py
 
 RUN chmod -R +x /usr/local/bin/generate* /usr/local/bin/set_irq_affinity /usr/local/bin/checkin_cli.py && \
   dpkg -i /var/tmp/python3-aioeventlet* && \
