+++ lte/gateway/python/magma/mobilityd/tests/BUILD.bazel
@@ -31,17 +31,6 @@ pytest_test(
 )
 
 pytest_test(
-    name = "test_dhcp_helper_cli",
-    size = "small",
-    srcs = ["test_dhcp_helper_cli.py"],
-    imports = [LTE_ROOT],
-    deps = [
-        "//lte/gateway/python/scripts:dhcp_helper_cli",
-        requirement("scapy"),
-    ],
-)
-
-pytest_test(
     name = "test_ip_allocator_dhcp",
     size = "small",
     srcs = ["test_ip_allocator_dhcp.py"],
